+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Added `craft\commerce\paypalcheckout\gateways\Gateway::setTestMode()`
 
 ### Fixed
-- Fixed a bug where “Test mode?” and “Landing Page” gateway settings were not being parsed as environment variables.
+- Fixed a bug where “Send Gateway Info”, “Test mode?” and “Landing Page” gateway settings were not being parsed as environment variables.
 
 ## 1.3.0 - 2021-12-01